+++ internal/pkg/warewulfd/util.go
@@ -45,9 +45,7 @@ func sendFile(
 }
 
 func getOverlayFile(n node.Node, context string, stage_overlays []string, autobuild bool) (stage_file string, err error) {
-
 	stage_file = overlay.OverlayImage(n.Id(), context, stage_overlays)
-	err = nil
 	build := !util.IsFile(stage_file)
 	wwlog.Verbose("stage file: %s", stage_file)
 	if !build && autobuild {
@@ -60,10 +58,23 @@ func getOverlayFile(n node.Node, context string, stage_overlays []string, autobu
 	}
 
 	if build {
+		registry, err := node.New()
+		if err != nil {
+			wwlog.Error("Failed to build overlay: %s, %s, %s\n%s",
+				n.Id(), stage_overlays, stage_file, err)
+			return "", err
+		}
+		var allNodes []node.Node
+		allNodes, err = registry.FindAllNodes()
+		if err != nil {
+			wwlog.Error("Failed to build overlay: %s, %s, %s\n%s",
+				n.Id(), stage_overlays, stage_file, err)
+			return "", err
+		}
 		if len(stage_overlays) > 0 {
-			err = overlay.BuildSpecificOverlays([]node.Node{n}, stage_overlays)
+			err = overlay.BuildSpecificOverlays([]node.Node{n}, allNodes, stage_overlays, 1)
 		} else {
-			err = overlay.BuildAllOverlays([]node.Node{n})
+			err = overlay.BuildAllOverlays([]node.Node{n}, allNodes, 1)
 		}
 		if err != nil {
 			wwlog.Error("Failed to build overlay: %s, %s, %s\n%s",